requirements.txt
strawberry-graphql==0.215.1
Flask==3.0.0
sqlalchemy==2.0.23